Fix implicit cast to int in dumper.php
> Unexpected PHP Error [Implicit conversion from float 4.5 to int loses
> precision] severity [8192] in [simpletest/dumper.php line 133]

+++ dumper.php
@@ -126,7 +126,7 @@ class SimpleDumper
             return $value;
         }
         $position = min($position, $length);
-        $start    = ($size/2 > $position ? 0 : $position - $size/2);
+        $start    = (int)($size/2 > $position ? 0 : $position - $size/2);
         if ($start + $size > $length) {
             $start = $length - $size;
         }
